Added selectors for terrain, territory, and other tags and updated tags container to use those instead of directly accessing state
